backend/pyproject.toml
[project]
name = "content-creator-workbench"
version = "0.1.0"
description = "Content Creator Workbench Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "google-api-python-client>=2.108.0",
    "google-auth>=2.23.4",
    "openai>=1.3.6",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.7",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"